+++ src/data/source/location.hpp
@@ -10,8 +10,8 @@ namespace amelia {
 struct Location {
   Text filename;
   CharIterator position;
-  size_t line;
-  size_t column;
+  uint32_t line;
+  uint32_t column;
 
   bool operator==(const Location &other) const {
     return filename == other.filename && line == other.line && column == other.column;
